refactor: Move Frequency based trip search out of raptor.
